Remove all references to prompting for user input and std::in
This was done because I could not figure out means of both
getting user input and storing it in a variable (but seriously,
why is this so difficult to do in node.js when it's a matter
of one line in languages like Python and Java?).

As a substitute, I've made all selection functions return the
same choice everytime.

Also added test suite for Alibi class.