Merge backend code using WebPack
Added preliminary code that demonstrates the use of webpack
with frontend user interaction. A user can control cards in
their hand by clicking a card to pick it up. Hovering over
a card will enlarge it. A user can click again with the card
in hand to return it to its original position.